+++ contracts/migrations/Migration.sol
@@ -5,13 +5,12 @@ import "../interfaces/IMigrationRegistry.sol";
 import "../interfaces/IHub.sol";
 import "../interfaces/IMeTokenRegistry.sol";
 
-
 contract Migration {
-
-    uint public immutable PRECISION = 10**18;
+    uint256 public immutable PRECISION = 10**18;
 
     // TODO
-    IMigrationRegistry public migrationRegistry = IMigrationRegistry(address(0));
+    IMigrationRegistry public migrationRegistry =
+        IMigrationRegistry(address(0));
     IHub public hub = IHub(address(0));
     IMeTokenRegistry public meTokenRegistry = IMeTokenRegistry(address(0));
 
@@ -20,5 +19,4 @@ contract Migration {
     function finish() public virtual {
         // send ending collateral to new hub
     }
-
 }
